Максимальный поток в сети
- Добавлен алгоритм Форда-Фалкерсона для нахождения максимального потока в сети
- Прочие небольшие исправления / улучшения

+++ NGProject/Pipe.h
@@ -14,7 +14,7 @@ public:
 	bool isInConnection = false;
 
 	int getID();
-	int getLength() const;
+	float getLength() const;
 	int getDiameter() const;
 	int getStatus() const;
